# coffee-app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Coffee Shop API",
        "description": "API for managing coffee shop operations",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/branches": {
            "get": {
                "tags": [
                    "Branches"
                ],
                "summary": "Get all branches",
                "operationId": "8bad73dccf3e88f9ee9007d8d3eb1040",
                "responses": {
                    "200": {
                        "description": "List of branches",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Branch"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Branches"
                ],
                "summary": "Create a new branch",
                "operationId": "df88a7dda440ba67d2d3433731a1dbee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "branch_name"
                                ],
                                "properties": {
                                    "branch_name": {
                                        "type": "string",
                                        "example": "New Branch"
                                    },
                                    "branch_address": {
                                        "type": "string",
                                        "example": "456 Elm St"
                                    },
                                    "branch_phone": {
                                        "type": "string",
                                        "example": "+0987654321"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Branch created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Branch"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/branches/{id}": {
            "get": {
                "tags": [
                    "Branches"
                ],
                "summary": "Get a specific branch",
                "operationId": "c9388b859d392f0ff3b7077e95d29776",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Branch details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Branch"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Branch not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Branches"
                ],
                "summary": "Update a branch",
                "operationId": "65ebd1d6e03a7404ecbdc887f49b5d08",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "branch_name": {
                                        "type": "string",
                                        "example": "Updated Branch"
                                    },
                                    "branch_address": {
                                        "type": "string",
                                        "example": "789 Oak St"
                                    },
                                    "branch_phone": {
                                        "type": "string",
                                        "example": "+1122334455"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Branch updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Branch"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Branch not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Branches"
                ],
                "summary": "Delete a branch",
                "operationId": "b4945d42e09ef04578cb6e61d86f156d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Branch deleted"
                    },
                    "404": {
                        "description": "Branch not found"
                    }
                }
            }
        },
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get all categories",
                "operationId": "3f5817a34833d0a1f4af4548dd3aeaba",
                "responses": {
                    "200": {
                        "description": "List of categories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Category"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Create a new category",
                "operationId": "71fcad552bb0eaba9fb191fd8d8dcab0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cat_name"
                                ],
                                "properties": {
                                    "cat_name": {
                                        "type": "string",
                                        "example": "Tea"
                                    },
                                    "cat_description": {
                                        "type": "string",
                                        "example": "Various tea options"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Category created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get a specific category",
                "operationId": "c68e76d323c008827a9e47398b1583de",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Update a category",
                "operationId": "0e686b2748211cc688333ed705dc111f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "cat_name": {
                                        "type": "string",
                                        "example": "Updated Tea"
                                    },
                                    "cat_description": {
                                        "type": "string",
                                        "example": "Updated description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Category updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Delete a category",
                "operationId": "4c12e22a7f8c617bd83bfa1fdc05428d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Category deleted"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Branch": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "branch_name": {
                        "type": "string",
                        "example": "Main Branch"
                    },
                    "branch_address": {
                        "type": "string",
                        "example": "123 Main St"
                    },
                    "branch_phone": {
                        "type": "string",
                        "example": "+1234567890"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Category": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "cat_name": {
                        "type": "string",
                        "example": "Coffee"
                    },
                    "cat_description": {
                        "type": "string",
                        "example": "Various coffee drinks"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}